Clean up the script usage texts
By using the ARGBASH_PREPARE directive instead of ARGBASH_GO, argbash does not
add a call to parse_commandline, which gives us an opportunity to override the
die and print_help functions, which lets us customize the command-line help to
whatever we would like.

print_option_usage was created to consolidate the help text of the CLI options
into one place, but still gives enough flexibility for the actual scripts to
include/exclude options as is appropriate for that script.