Declare additional inputs and outputs for generateBashCliParsers task
The task should be re-run if one of the .m4 template files change.

The input scripts of the task are also its outputs. This can lead to a case
where the build has to be run twice before the up-to-date check will pass
(particularly if a m4 template file was changed), but this is still an
improvement.