Eliminate publish tasks (for now)
We can eliminate the publish tasks by setting the directory distributions are
saved in to the final folder we want the zips to land in.

This also eliminates the warnings from Gradle about the assemble tasks using
outputs from the publish tasks (which is a side-effect of having the publish
tasks write their output to the root directory).

We can re-add publish tasks later when we find a better place to publish the
actual archive files.