Fix bug: capturing the 2nd build scan sometimes fails with "could not find method"
It's not entirely clear why, but sometimes the buildScanPublished Action is
unable to directly resolve references to methods and variables within the init
script. When this happens, the build scan is not logged and a warning is printed to standard out:

```
WARNING: Error invoking build scan buildScanPublished action
  Could not find method logScan() for arguments [com.gradle.scan.plugin.internal.o.n$1@23035cca] on object of type com.gradle.scan.plugin.internal.o.n$1.
```

Storing a reference to the logBuildScan function and then using the
reference in the closure seems to work around the bug.