Update Gradle experiment 03 to copy the project before running any builds
This allows the script to literally copy the project. Since no builds have
executed yet, there is no build output (or other possible changes) that have to
be reversed.

A major advantage of this approach is that the build scans produced by both
builds will contain the same data about the git repository. For example, if the
project uses the Common Custom User Data plugin, then the Git repository custom
value will be the same for both builds (previously, the second build's
repository would actually point back to the original cloned project).