Capture the Maven project name like the GE Maven Extension
The previous method "worked", but it is unclear how robust it was (would it
work with a wide range of Maven versions, and with a wide range of different
types of Maven projects?).

The new method, which was taken from the Gradle Enterprise Maven Extension is
likely to be more robust because the Gradle Enterprise Maven Extension has been
heavily tested (both in-house and by actual customers).

This also has the advantage of ensuring the project name extracted here matches
the project name that is in the build scans (because we get it the same way).