Print simple summary after copying a project
Print a summary after copying files in a portable way is difficult in bash.
Most guidance is to use an external command that may or may not be on our
target user's systems. One suggestion is to use rsync, but the old version of
rsync that ships with MacOS doesn't have support for the kind of simple output
we are looking for.

It is safe to assume the wc command is generally available, so using it, and
cp's -v flag, it's possible to at least report on the number of files copied.