Fix experiment dir names
The RUN_ID has to be initialized before the EXP_DIR constant.

Originally, I had moved RUN_ID out of the init.sh so that we could define it
conditionally (exp4-gradle and exp3-maen do not need it), but that lead to this
bug. Instead, a new constant is defined for every experiment to state whether
or not to include the run ID in the summary.