Remove the clean task from tasks detected in build scans
It is easier and cleaner to just exclude clean from the tasks fetched from all
build scans than to try to exclude it later, conditionally, for Gradle exp5 and
Maven exp4. There is not a significant downside either: this way, the data in
the task comparison is always consistent.

It is better to do this when we're processing the data fetched from the build
scans because we compare the tasks in several places.