Mark generateSwaggerCode tasks as not being configuration cache compatible.
This allows the build to work, but the configuration cache is used a lot less
as a result. The better solution is to ask the codegen maintainers to fix the
plugin to be configuraiton-cache compatible.

diff --git a/components/fetch-build-scan-data-cmdline-tool/build.gradle.kts
b/components/fetch-build-scan-data-cmdline-tool/build.gradle.kts index
068db94..f93bcb4 100644 ---
a/components/fetch-build-scan-data-cmdline-tool/build.gradle.kts +++
b/components/fetch-build-scan-data-cmdline-tool/build.gradle.kts @@ -1,3 +1,5
@@ +import org.hidetake.gradle.swagger.generator.GenerateSwaggerCode + plugins
{ java application @@ -65,6 +67,10 @@ tasks.compileJava { options.encoding =
"UTF-8" }

+tasks.withType(GenerateSwaggerCode::class) {
+    notCompatibleWithConfigurationCache("Accesses the project at execution time.")
+}
+
 application {
     mainClass.set("com.gradle.enterprise.Main")
 }