Switch to the OpenAPI Generator 6.0.0 Gradle Plugin
This replaces the old Swagger Generator Gradle plugin. It has the following
benefits:

	* More closely follows normal Gradle plugin conventions (for example, there
	  isn't the weird swaggerSources container anymore, and there are instead more normal convention objects for the tasks the plugin provides).
  * Is fully incremental building, build caching, and configuration-cache compatible.

This commit also switches to using an .openapi-generator-ignore file to exclude
the files we don't need to have generated. The file has to be at the root of
the module otherwise it does not appear to work correctly, and the entries
within the file must be relative to the project root.