Only allow untrusted SSL certificates for the Gradle Enterprise host
This is a little more secure. If, in the future,
fetch-build-scan-data-cmdline-tool connects to other hosts, then it will still
validate their SSL certificates. It will also validate that the hostnames on
those certificates match the requested hostnames (HostnameVerifiers are only
called if the hostnames do not match the hosts on the certificates).

This is also more consistent with the Gradle Enterprise Gradle Plugin and Maven
extension.

To avoid confusion, this commit also moves away from using javax.net.ssl as a
prefix for our custom property. Using the prefix made it look like it was a
built-in java SSL system property, which it is not.