Make build scan data fetching work with GE 2022.1 - 2022.2.3 again
If build scan data is fetched from a version of GE prior to 2022.2.4, then
executed non-cacheable tasks with overlapping inputs or for which caching was
disabled to ensure correctness are counted as executed non-cacheable tasks
(instead of executed cacheable tasks).

The taskType property (which was introduced to the GE API in 2022.2.4) needed
to be made optional in the OpenAPI spec in order for the OpenAPI generator to
create a client that works with GE 2022.1 - 2022.2.3 (otherwise fetching build
scan data would fail with "missing required property taskType").