gradle exp 05/maven exp 04: Fix the build caching stats to be based on the 2nd build
Previously, the arrays containing the build scan data were built using append
operations.  Since Gradle Exp 05 and Maven Exp 04 fetch build scans twice (once
to get the build parameters for the 2nd build), this means the data for the
first build was appended to the array twice. When it came time to display the
statistics, the scripts always show the data at array index 1 (zero-based
indexing), so the statistics for the first build would be displayed.