Fix the -x shorthand for --no-build-scan-publishing
This is a workaround for a limitation with how Argbash handles boolean CLI
switches. Basically, there is no way to set a shorthand for a boolean CLI
switch to set the its value to "off".

The actual --disable-build-scan-publishing CLI argument won't be presented to
users, but if a user happens to use it, it will do the right thing. There is
some potentially odd behavior if both flags are used at the same time, but this
seems unlikely.