Do not require the user to provide a GE server if Build Scan publishing is disabled
If Build Scan publishing is disabled, then the GE server used doesn't matter.
It isn't captured in the Build Scan dumps, and even if it was, it wouldn't
affect the experiment.

As such, if the user provides `--enable-gradle-enterprise` along with
`--disable-build-scan-publishing`, then no longer require the user to also
provide `--gradle-enterprise-server`.

Unfortunately, the Gradle Enterprise plugin and Maven extension still require a
server to run, so we want to set the server to a dummy value, but only if the
user isn't setting `--gradle-enterprise-server` themselves.