Show the command to repeat the experiment early on in interactive mode
If the experiment needs to be interrupted for some reason, it's rather annoying
to have to re-enter all of the settings when re-running the experiment. We show
the command to repeat the experiment at the end of the summary, but that's only
helpful if the script completes normally.

This commit updates interactive mode to also show the command to repeat the
experiment immediately after all of the configuration options have been
collected, but before executing any of the builds.