Better handle cases where builds fail because the JVM exits unexpectedly
Previously, the build validation scripts depended on the order build scans were
written to the build_scans.csv file, but this led to subtle and weird bugs,
particularly if the JVM running the build exits unexpectedly (such as in
response to an OutOfMemoryError).

This commit reworks things so that each row in the build_scans.csv file is
directly associated with the experiment build that generated it. When reading
build scan data, the association is re-established by matching build scan ids
with those that were produced by the builds.

A few other changes were made so that the build summary output makes sense,
particularly if the first build fails because the JVM exits unexpectedly.