Better handle cases where builds fail because the JVM exits unexpectedly (#334)
Previously, the build validation scripts depended on the order build scans were
written to the build_scans.csv file, but this led to subtle and weird bugs,
particularly if the JVM running the build exits unexpectedly (such as in
response to an OutOfMemoryError).

This commit reworks things so that each row in the build_scans.csv file is
directly associated with the experiment build that generated it. When reading
build scan data, the association is re-established by matching build scan ids
with those that were produced by the builds.

For Gradle Exp 05/Maven Exp 04, we need to have the build scan ID to 
correctly associate the Build Scan data that is fetched, even if a build 
isn't executed. This was done for Gradle Exp 04 / Maven Exp 03, but not 
for the later experiment. As a bonus, this PR deduplicates the 
Build Scan URL parsing logic.

A few other changes were made so that the build summary output makes sense,
particularly if the first build fails because the JVM exits unexpectedly.