Fetch Build Scan data in parallel
Some users have builds which produce a lot of Build Scan data, and so it can
sometimes take a while to fetch the data for all of the builds. Fetching the
Build Scan data in parallel provides a nice performance boost for such users.

In order to do this safely, the logging format needed to be adjusted and made
stateless (before we were tracking some state to provide a more concise log
output format).