Delay GE server URL check for Gradle >=6
Now that we check in the shell scripts for the presence of the GE server url when instrumenting the build with GE, we should rarely get to the scenario where there is no server GE URL specified by the time the Gradle init script is invoked. In the case it happens, a buildFinished warning caused by the CCUD plugin will be shown on the cmd line but has no negative effect.

This change makes it easier to reason about the code flow for different Gradle versions.