Remove unused variables and suppress deprecation warnings
Unused warnings were previously suppressed because the main classes aren't used
directly. However, this suppressed other unused warnings and hid a few things
that are not actually used.  So this commit unsupresses unused warnings and
removes the unused variables and parameters.

The attempt to suppress deprecation warnigns on the ConfigureGradleEnterprise
class wasn't working because the deprecated classes are referenced in import
statements. The class was updated to use the fully qualified name of the
deprecated classes.