Add support for the renamed Develocity Maven Extension (#624)
The Gradle Enterprise Maven Extension was renamed to the Develocity Maven
Extension in version 1.21.  The API was also updated for the product name
change. Unfortunately, this broke compatibility with the Build Validation
Scripts. This commit fixes the incompatibility.

The fix works by updating the configure-gradle-enterprise-maven-extension to
define and register two listeners: a GradleEnterpriseListener and a
DevelocityListener.  If an older version of the Gradle Enterprise Maven
Extension is used, it will load the GradleEnterpriseListener.  If a newer
version of the Develocity Maven Extension is used, then it will load the
DevelocityListener. The Develocity Maven Extension Adapters are used to
eliminate duplication between the two different listeners.

This commit also creates a "fat jar" so that the develocity-maven-extention-adapters jar
does not have to be manually added to the classpath.

The project compiles agianst an older version of Maven to ensure compatibility
with projects using an older version of Maven.

**Remove unused variables and suppress deprecation warnings**

Unused warnings were previously suppressed because the main classes aren't used
directly. However, this suppressed other unused warnings and hid a few things
that are not actually used.  So this commit un-supresses unused warnings and
removes the unused variables and parameters.

The attempt to suppress deprecation warnings on the ConfigureGradleEnterprise
class wasn't working because the deprecated classes are referenced in import
statements. The class was updated to use the fully qualified name of the
deprecated classes.

---------

Signed-off-by: Jim Hurne <jhurne@gradle.com>
Co-authored-by: Eric Haag <ehaag@gradle.com>
Co-authored-by: Brian Demers <brian.demers@gmail.com>